ROX-26854: move postgres flowstore tests to subdir
