+++ central/networkgraph/flow/datastore/internal/store/postgres/tests/flow_impl_test.go
@@ -1,10 +1,11 @@
 //go:build sql_integration
 
-package postgres
+package tests
 
 import (
 	"testing"
 
+	postgresFlowStore "github.com/stackrox/rox/central/networkgraph/flow/datastore/internal/store/postgres"
 	"github.com/stackrox/rox/central/networkgraph/flow/datastore/internal/store/testcommon"
 	"github.com/stackrox/rox/pkg/postgres/pgtest"
 	"github.com/stretchr/testify/suite"
@@ -14,7 +15,7 @@ func TestFlowStore(t *testing.T) {
 	testDB := pgtest.ForT(t)
 	defer testDB.DB.Close()
 
-	store := NewClusterStore(testDB.DB)
+	store := postgresFlowStore.NewClusterStore(testDB.DB)
 	flowSuite := testcommon.NewFlowStoreTest(store)
 	suite.Run(t, flowSuite)
 }
